requirements.txt
mkdocs==1.6.1
mkdocs-material==9.6.12
mkdocs-awesome-pages-plugin==2.10.1
mkdocs-macros-plugin==1.5.0
pyyaml==6.0.3
